% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDboost-package.R
\docType{package}
\name{FDboost-package}
\alias{FDboost-package}
\alias{_PACKAGE}
\alias{FDboost_package}
\alias{package-FDboost}
\title{FDboost: Boosting Functional Regression Models}
\description{
Regression models for functional data, i.e., scalar-on-function, 
function-on-scalar and function-on-function regression models, are fitted 
by a component-wise gradient boosting algorithm.
}
\details{
This package is intended to fit regression models with functional variables. 
It is possible to fit models with functional response and/or functional covariates, 
resulting in Scalar-on-function, function-on-scalar and function-on-function regression. 
Details on the functional regression models that can be fitted with \pkg{FDboost}
can be found in Brockhaus et al. (2015, 2016, 2017) and Ruegamer et al. (2016).  

For model fitting by component-wise gradient boosting, \pkg{FDboost} relies on 
the R package \pkg{mboost} (Hothorn et al., 2017). 
A comprehensive tutorial to \pkg{mboost} is given in Hofner et al. (2014). 

The main fitting function is \code{\link{FDboost}}. 
Like the fitting procedures in \pkg{mboost}, the function \code{FDboost} DOES NOT 
automatically estimate an appropriate model complexity. The most important tuning parameter
is the number of boosting iterations, which must be chosen by the user.  
It is recommended to determine an adequate stopping iteration by resampling methods like 
cross-validation or bootstrapping. 
This can be done using the function \code{\link{applyFolds}}.
}
\references{
Brockhaus, S., Scheipl, F., Hothorn, T. and Greven, S. (2015): 
The functional linear array model. Statistical Modelling, 15(3), 279-300. 

Brockhaus, S., Melcher, M., Leisch, F. and Greven, S. (2017): 
Boosting flexible functional regression models with a high number of functional historical effects,  
Statistics and Computing, 27(4), 913-926.   

Brockhaus, S., Fuest, A., Mayr, A. and Greven, S. (2016): 
Signal regression models for location, scale and shape with an application to stock returns. 
\url{https://arxiv.org/abs/1605.04281}

Hothorn T., Buehlmann P., Kneib T., Schmid M., and Hofner B. (2017). mboost: Model-Based Boosting, 
R package version 2.8-0, \url{https://cran.r-project.org/package=mboost}

Hofner, B., Mayr, A., Robinzonov, N., Schmid, M. (2014). Model-based Boosting in R: 
A Hands-on Tutorial Using the R Package mboost. Computational Statistics, 29, 3-35. 
\url{https://cran.r-project.org/package=mboost/vignettes/mboost_tutorial.pdf}

Ruegamer D., Brockhaus, S., Gentsch K., Scherer, K., Greven, S. (2016). Detecting synchronisation in 
EEG- and EMG-Signals via boosted functional historical models. 
\url{http://arxiv.org/abs/1609.06070}
}
\seealso{
\code{\link{FDboost}} for the main fitting function and 
\code{\link{applyFolds}} for model tuning via resampling methods.
}
\author{
Sarah Brockhaus and David Ruegamer
}
