\name{integrationWeights}
\alias{integrationWeights}
\title{Function to compute integration weights}
\usage{
  integrationWeights(X1, xind)
}
\arguments{
  \item{X1}{matrix of functional variable}

  \item{xind}{index of functional variable}
}
\description{
  Computes trapezoidal integration weights for a functional
  variable X1 on grid xind.
}

