\name{o_control}
\alias{o_control}
\title{Function to control estimation of smooth offset}
\usage{
  o_control(k_min = 20, silent = TRUE)
}
\arguments{
  \item{k_min}{maximal number of k in s()}

  \item{silent}{print error messages of model fit?}
}
\description{
  Function to control estimation of smooth offset
}

