% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCIs.R
\name{plot.bootstrapCI}
\alias{plot.bootstrapCI}
\title{Plot a bootstrapCI object}
\usage{
\method{plot}{bootstrapCI}(x, ...)
}
\arguments{
\item{x}{a fitted \code{bootstrapCI}-object}

\item{...}{ignored}
}
\value{
a list of \code{ggplot} objects
}
\description{
For scalar non-functional covariates, a boxplot is 
used to visualize the distribution. Quantiles are drawn by 
vertical red lines. For scalar response, functional covariates
are visualized by line plots with quantiles in dashed red.
For function-on-function effect surfaces, the quantiles are plotted
as surface plots.
}
\details{
Takes a \code{bootstrapCI}-object and produces ggplot objects
}
\note{
Quantiles are plotted using \code{ggplot}.
}
\seealso{
\code{\link{bootstrapCI}}
}

