% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{predict.FDboost_fac}
\alias{predict.FDboost_fac}
\alias{plot.FDboost_fac}
\title{Prediction and plotting for factorized FDboost model components}
\usage{
\method{predict}{FDboost_fac}(object, newdata = NULL, which = NULL, ...)

\method{plot}{FDboost_fac}(x, which = NULL, main = NULL, ...)
}
\arguments{
\item{object, x}{a model-factor given as a \code{FDboost_fac} object}

\item{newdata}{optionally, a data frame or list
in which to look for variables with which to predict.
See \code{\link{predict.mboost}}.}

\item{which}{a subset of base-learner components to take into
account for computing predictions or coefficients. Different
components are never aggregated to a joint prediction, but always
returned as a matrix or list. Select the k-th component
by name in the format \code{bl(x, ...)[k]} or all components of a base-learner
by dropping the index or all base-learners of a variable by using
the variable name.}

\item{...}{additional arguments passed to underlying methods.}

\item{main}{the plot title. By default, base-learner names are used with 
component numbers \code{[k]}.}
}
\description{
Prediction and plotting for factorized FDboost model components
}
\seealso{
[factorize(), factorize.FDboost()]
}
