% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCIs.R
\name{print.bootstrapCI}
\alias{print.bootstrapCI}
\title{Print a bootstrapCI object}
\usage{
\method{print}{bootstrapCI}(x, ...)
}
\arguments{
\item{x}{a fitted \code{bootstrapCI}-object}

\item{...}{currently not used}
}
\value{
a list with information on the model
}
\description{
Takes a \code{bootstrapCI}-object and produces a print on the console.
}
\seealso{
\code{\link{bootstrapCI}}
}

