% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.FDboost}
\alias{print.FDboost}
\title{Print a boosted functional regression model}
\usage{
\method{print}{FDboost}(x, ...)
}
\arguments{
\item{x}{a fitted \code{FDboost}-object}

\item{...}{currently not used}
}
\value{
a list with information on the model
}
\description{
Takes a fitted \code{FDboost}-object and produces a print on the console.
}
\seealso{
\code{\link{FDboost}} for the model fit.
}

