% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.FDboost}
\alias{summary.FDboost}
\title{Summary of a boosted functional regression model}
\usage{
\method{summary}{FDboost}(object, ...)
}
\arguments{
\item{object}{a fitted \code{FDboost}-object}

\item{...}{currently not used}
}
\value{
a list with summary information
}
\description{
Takes a fitted \code{FDboost}-object and produces a summary.
}
\seealso{
\code{\link{FDboost}} for the model fit.
}

